refactor: split conformance report logic from CLI
